Moving constants to separate fiel
